# Conflicts:
#	Assets/Marble Effects.meta
#	Assets/Marble Effects/ Marble Falls in Glass Hole 1.wav.meta
#	Assets/Marble Effects/ Marble Falls in Glass Hole 2.wav.meta
#	Assets/Marble Effects/ Marble Falls in Glass Hole 3.wav.meta
#	Assets/Marble Effects/ Marble Hits Cloth.wav.meta
#	Assets/Marble Effects/ Marble Hits Glass 1.wav.meta
#	Assets/Marble Effects/ Marble Hits Glass 2.wav.meta
#	Assets/Marble Effects/ Marble Hits Glass 3.wav.meta
#	Assets/Marble Effects/ Marble Hits Glass 4.wav.meta
#	Assets/Marble Effects/ Marble Hits Metal 1.wav.meta
#	Assets/Marble Effects/ Marble Rolls on Glass 1.wav.meta
#	Assets/Marble Effects/ Marble Rolls on Glass 2.wav.meta
#	Assets/Marble Effects/ Marble Rolls on Glass 3.wav.meta
#	Assets/Marble Effects/ Marble Rolls on Glass 4.wav.meta
#	Assets/Marble Effects/ Marble Rolls on Glass 5.wav.meta
#	Assets/Scenes/main_menu.unity

+++ Assets/Scripts/MainMenu/MainMenu.cs
@@ -65,7 +65,7 @@ public class MainMenu : MonoBehaviour
 		JerkPreferences jerkData = new JerkPreferences();
 		jerkData.jerkCount = (int)GameObject.Find("JerkCountSlider").GetComponent<Slider>().value;
 		jerkData.jerkWeight = GameObject.Find("JerkWeightSlider").GetComponent<Slider>().value;
-		jerkData.jerkInterval = (int)GameObject.Find("JerkIntervalSlider").GetComponent<Slider>().value;
+		jerkData.jerkInterval = GameObject.Find("JerkIntervalSlider").GetComponent<Slider>().value;
 		
 		DataSaver.saveData(jerkData, "Jerks");
